FINALLY VOLATILE CONTRAPTION REWORK IS DONE

+++ Content/Projectiles/Probe.cs
@@ -1,3 +1,4 @@
+using DartGunsPlus.Content.Dusts;
 using DartGunsPlus.Content.Systems;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
@@ -47,7 +48,12 @@ public class Probe : ModProjectile
         Projectile.rotation = Projectile.DirectionTo(Target.Center).ToRotation();
 
         if (!Target.active)
-            Projectile.Kill();
+        {
+            if (Projectile.FindTargetWithinRange(2000) is not null)
+                Projectile.ai[0] = Projectile.FindTargetWithinRange(2000).whoAmI; // target
+            else
+                Projectile.Kill();
+        }
         
         const int dist = 150;
         
@@ -60,19 +66,19 @@ public class Probe : ModProjectile
         {
             Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.Center, Projectile.DirectionTo(Target.Center) * 6,
                 ModContent.ProjectileType<VolatileBolt>(), Projectile.damage, 2, Projectile.owner);
-            VisualSystem.SpawnDustCircle(Projectile.Center, DustID.RainbowRod, 12, color: Color.HotPink, scale: 0.9f);
+            VisualSystem.SpawnDustCircle(Projectile.Center, ModContent.DustType<GlowFastDecelerate>(), 12, color: Color.Red, scale: 0.9f);
             SoundEngine.PlaySound(AudioSystem.ReturnSound("shoot1"), Projectile.Center);
             
             Projectile.velocity = Projectile.DirectionFrom(Target.Center) * 6;
             RecoilTimer++;
-            RotationGoal *= Main.rand.Next(-3, -1);
+            RotationGoal *= -1;
         }
 
         if (RecoilTimer == 0)
         {
             Projectile.Center = Vector2.Lerp(Projectile.Center, Target.Center + new Vector2(dist, 0).RotatedBy(RotationOffset), 0.04f);
             Projectile.velocity = Vector2.Zero;
-            RotationOffset = (float)Utils.Lerp(RotationOffset, RotationGoal, 0.01f);
+            RotationOffset = (float)Utils.Lerp(RotationOffset, RotationGoal, 0.04f);
         }
     }
